+++ CHANGELOG.md
@@ -3,9 +3,14 @@
 ## Unreleased
 
 ### New Features
+- Added support for Azure Cognitive Search vector store (#7469)
+
+## [0.8.20] - 2023-09-04
+
+### New Features
 - Added Portkey LLM integration (#7508)
 - Support postgres/pgvector hybrid search (#7501)
-- Added support for Azure Cognitive Search vector store (#7469)
+- upgrade recursive retriever node reference notebook (#7537)
 
 ## [0.8.19] - 2023-09-03